requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.10.2
python-dotenv==1.0.1

requests==2.32.3
httpx==0.27.2

spacy==3.8.2
openai==1.57.3

mlflow==3.7.0
